move domain classes for referendumvorlagen to subpackage
