+++ wls-basisdaten-service/src/test/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/services/referendumvorlage/ReferendumvorlageModelMapperTest.java
@@ -1,8 +1,8 @@
 package de.muenchen.oss.wahllokalsystem.basisdatenservice.services.referendumvorlage;
 
-import de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.Referendumoption;
-import de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.Referendumvorlage;
-import de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.Referendumvorlagen;
+import de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.referendumvorlagen.Referendumoption;
+import de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.referendumvorlagen.Referendumvorlage;
+import de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.referendumvorlagen.Referendumvorlagen;
 import de.muenchen.oss.wahllokalsystem.wls.common.security.domain.BezirkUndWahlID;
 import java.util.Set;
 import java.util.UUID;
@@ -43,7 +43,7 @@ class ReferendumvorlageModelMapperTest {
 
             val entityToMap = new Referendumvorlagen(UUID.randomUUID(), new BezirkUndWahlID(wahlID, wahlbezirkID), stimmzettelgebietID,
                     Set.of(new Referendumvorlage(null, null, "wahlvorschlagID1", 1L, "kurzname1", "frage1",
-                            Set.of(new Referendumoption("option11", "optionsName11", 1L), new Referendumoption("option12", "optionsName12", 2L))),
+                                    Set.of(new Referendumoption("option11", "optionsName11", 1L), new Referendumoption("option12", "optionsName12", 2L))),
                             new Referendumvorlage(null, null, "wahlvorschlagID2", 2L, "kurzname2", "frage2",
                                     Set.of(new Referendumoption("option21", "optionsName21", 3L), new Referendumoption("option22", "optionsName22", 4L)))));
 
@@ -51,8 +51,8 @@ class ReferendumvorlageModelMapperTest {
 
             val expectedResult = new ReferendumvorlagenModel(stimmzettelgebietID,
                     Set.of(new ReferendumvorlageModel("wahlvorschlagID1", 1L, "kurzname1", "frage1",
-                            Set.of(new ReferendumoptionModel("option11", "optionsName11", 1L),
-                                    new ReferendumoptionModel("option12", "optionsName12", 2L))),
+                                    Set.of(new ReferendumoptionModel("option11", "optionsName11", 1L),
+                                            new ReferendumoptionModel("option12", "optionsName12", 2L))),
                             new ReferendumvorlageModel("wahlvorschlagID2", 2L, "kurzname2", "frage2",
                                     Set.of(new ReferendumoptionModel("option21", "optionsName21", 3L),
                                             new ReferendumoptionModel("option22", "optionsName22", 4L)))));
@@ -73,8 +73,8 @@ class ReferendumvorlageModelMapperTest {
 
                 val modelToMap = new ReferendumvorlagenModel(stimmzettelgebietID,
                         Set.of(new ReferendumvorlageModel("wahlvorschlagID1", 1L, "kurzname1", "frage1",
-                                Set.of(new ReferendumoptionModel("option11", "optionsName11", 1L),
-                                        new ReferendumoptionModel("option12", "optionsName12", 2L))),
+                                        Set.of(new ReferendumoptionModel("option11", "optionsName11", 1L),
+                                                new ReferendumoptionModel("option12", "optionsName12", 2L))),
                                 new ReferendumvorlageModel("wahlvorschlagID2", 2L, "kurzname2", "frage2",
                                         Set.of(new ReferendumoptionModel("option21", "optionsName21", 3L),
                                                 new ReferendumoptionModel("option22", "optionsName22", 4L)))));
@@ -84,7 +84,7 @@ class ReferendumvorlageModelMapperTest {
 
                 val expectedResult = new Referendumvorlagen(null, bezirkUndWahlID, stimmzettelgebietID,
                         Set.of(new Referendumvorlage(null, null, "wahlvorschlagID1", 1L, "kurzname1", "frage1",
-                                Set.of(new Referendumoption("option11", "optionsName11", 1L), new Referendumoption("option12", "optionsName12", 2L))),
+                                        Set.of(new Referendumoption("option11", "optionsName11", 1L), new Referendumoption("option12", "optionsName12", 2L))),
                                 new Referendumvorlage(null, null, "wahlvorschlagID2", 2L, "kurzname2", "frage2",
                                         Set.of(new Referendumoption("option21", "optionsName21", 3L), new Referendumoption("option22", "optionsName22", 4L)))));
                 Assertions.assertThat(result).isEqualTo(expectedResult);
